(maint) Update gemfile and gemspec
This commit updates the gemfile and gemspec in prepartion for the actual code
to do the module building.  This is based on the versions in the PDK which is
where this code originated from. This commit also adds the Gemfile.local bundle
pattern to make it easier for people to add their own preferred debugging gems
and tools.