Preserve directory mtimes
To make module building more idempotent, this copies directory mtimes
after copying files. The result is that when removing the staging
directory and creating another the mtimes will be the same.

Since creating or removing a file modifies a directory, it needs to
happen after all files have been copied.

This can be beneficial when using rsync to deploy.