Add a setter for release_name
I would like to be able to build to target/modules/apache instead of
target/modules/puppetlabs-apache-1.2.3. I can achieve target/modules
instead of ./pkg via the destination constructor argument, but the
release name is not configurable. It can only be set via an instance
variable. This setter adds a formal API.